fix: address automated review standards for export PR
Merge error-handling tests into useLazyFetchAllRecords.test.tsx using
the standard describe('error handling') pattern with await act().

Split the RLS predicate util into focused helpers so the main file stays
under the 500-line service limit.
