+++ packages/twenty-front/src/modules/object-record/hooks/__tests__/useLazyFetchAllRecords.test.tsx
@@ -1,174 +1,145 @@
-import { PERSON_FRAGMENT_WITH_DEPTH_ZERO_RELATIONS } from '@/object-record/hooks/__mocks__/personFragments';
+/* @license Enterprise */
+
+import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
 import { useLazyFetchAllRecords } from '@/object-record/hooks/useLazyFetchAllRecords';
-import { type MockedResponse } from '@apollo/client/testing';
+import { useLazyFindManyRecords } from '@/object-record/hooks/useLazyFindManyRecords';
 import { act, renderHook, waitFor } from '@testing-library/react';
-import gql from 'graphql-tag';
-import { getJestMetadataAndApolloMocksAndCommandMenuWrapper } from '~/testing/jest/getJestMetadataAndApolloMocksAndCommandMenuWrapper';
-import { getTestEnrichedObjectMetadataItemsMock } from '~/testing/utils/getTestEnrichedObjectMetadataItemsMock';
 
-const defaultResponseData = {
-  pageInfo: {
-    hasNextPage: false,
-    hasPreviousPage: false,
-    startCursor: '',
-    endCursor: '',
-  },
-  totalCount: 2,
-};
+jest.mock('@/object-metadata/hooks/useObjectMetadataItem');
+jest.mock('@/object-record/hooks/useLazyFindManyRecords');
 
-const mockPerson = {
-  __typename: 'Person',
-  updatedAt: '2021-08-03T19:20:06.000Z',
-  whatsapp: {
-    primaryPhoneNumber: '+1',
-    primaryPhoneCountryCode: '234-567-890',
-    additionalPhones: [],
-  },
-  linkedinLink: {
-    primaryLinkUrl: 'https://www.linkedin.com',
-    primaryLinkLabel: 'linkedin',
-    secondaryLinks: ['https://www.linkedin.com'],
-  },
-  name: {
-    firstName: 'firstName',
-    lastName: 'lastName',
-  },
-  emails: {
-    primaryEmail: 'email',
-    additionalEmails: [],
-  },
-  position: 'position',
-  createdBy: {
-    source: 'source',
-    workspaceMemberId: '1',
-    name: 'name',
-  },
-  avatarUrl: 'avatarUrl',
-  jobTitle: 'jobTitle',
-  xLink: {
-    primaryLinkUrl: 'https://www.linkedin.com',
-    primaryLinkLabel: 'linkedin',
-    secondaryLinks: ['https://www.linkedin.com'],
-  },
-  performanceRating: 1,
-  createdAt: '2021-08-03T19:20:06.000Z',
-  phones: {
-    primaryPhoneNumber: '+1',
-    primaryPhoneCountryCode: '234-567-890',
-    additionalPhones: [],
-  },
-  id: '123',
-  city: 'city',
-  companyId: '1',
-  intro: 'intro',
-  deletedAt: null,
-  workPreference: 'workPreference',
-};
-
-const mock: MockedResponse = {
-  request: {
-    query: gql`
-          query FindManyPeople(
-              $filter: PersonFilterInput
-              $orderBy: [PersonOrderByInput]
-              $lastCursor: String
-              $limit: Int
-              $offset: Int
-          ) {
-              people(
-                  filter: $filter
-                  orderBy: $orderBy
-                  first: $limit
-                  after: $lastCursor
-                  offset: $offset
-              ) {
-                  edges {
-                      node {
-                          ${PERSON_FRAGMENT_WITH_DEPTH_ZERO_RELATIONS}
-                      }
-                      cursor
-                  }
-                  pageInfo {
-                      hasNextPage
-                      hasPreviousPage
-                      startCursor
-                      endCursor
-                  }
-                  totalCount
-              }
-          }
-      `,
-    variables: {
-      limit: 30,
-    },
-  },
-  result: jest.fn(() => ({
-    data: {
-      people: {
-        ...defaultResponseData,
-        edges: [
-          {
-            node: mockPerson,
-            cursor: '1',
-          },
-          {
-            node: mockPerson,
-            cursor: '2',
-          },
-        ],
-      },
-    },
-  })),
-};
-
-const Wrapper = getJestMetadataAndApolloMocksAndCommandMenuWrapper({
-  apolloMocks: [mock],
-  componentInstanceId: 'recordIndexId',
-  contextStoreTargetedRecordsRule: {
-    mode: 'selection',
-    selectedRecordIds: [],
-  },
-  contextStoreCurrentObjectMetadataNameSingular: 'person',
-});
+const mockUseObjectMetadataItem = jest.mocked(useObjectMetadataItem);
+const mockUseLazyFindManyRecords = jest.mocked(useLazyFindManyRecords);
 
 describe('useLazyFetchAllRecords', () => {
-  const objectNameSingular = 'person';
-  const objectMetadataItem = getTestEnrichedObjectMetadataItemsMock().find(
-    (item) => item.nameSingular === objectNameSingular,
-  );
-  if (!objectMetadataItem) {
-    throw new Error('Object metadata item not found');
-  }
+  const mockFindManyRecordsLazy = jest.fn();
+  const mockFetchMoreRecordsLazy = jest.fn();
+
+  beforeEach(() => {
+    jest.clearAllMocks();
+    mockUseObjectMetadataItem.mockReturnValue({
+      objectMetadataItem: {
+        nameSingular: 'person',
+        namePlural: 'people',
+      } as any,
+    });
+    mockUseLazyFindManyRecords.mockReturnValue({
+      findManyRecordsLazy: mockFindManyRecordsLazy,
+      fetchMoreRecordsLazy: mockFetchMoreRecordsLazy,
+      queryIdentifier: 'test-query',
+    });
+  });
 
   it('should handle one single page', async () => {
-    const { result } = renderHook(
-      () =>
-        useLazyFetchAllRecords({
-          objectNameSingular,
-          limit: 30,
-        }),
-      {
-        wrapper: Wrapper,
+    const mockPerson = { id: '1' };
+
+    mockFindManyRecordsLazy.mockResolvedValue({
+      data: {
+        people: {
+          edges: [{ node: mockPerson, cursor: '1' }],
+          pageInfo: {
+            hasNextPage: false,
+            hasPreviousPage: false,
+            startCursor: '1',
+            endCursor: '1',
+          },
+          totalCount: 1,
+        },
       },
+      error: undefined,
+    });
+
+    const { result } = renderHook(() =>
+      useLazyFetchAllRecords({
+        objectNameSingular: 'person',
+        limit: 30,
+      }),
     );
 
-    let res: any;
+    let fetchPromise: Promise<unknown>;
 
     act(() => {
-      res = result.current.fetchAllRecords();
+      fetchPromise = result.current.fetchAllRecords();
     });
 
     expect(result.current.isDownloading).toBe(true);
+
+    const finalResult = await fetchPromise!;
 
     await waitFor(() => {
       expect(result.current.isDownloading).toBe(false);
       expect(result.current.progress).toEqual({ displayType: 'number' });
     });
 
-    expect(result.current.progress).toEqual({ displayType: 'number' });
+    expect(finalResult).toEqual([mockPerson]);
+  });
 
-    const finalResult = await res;
+  describe('error handling', () => {
+    it('should reject when a subsequent page fetch fails', async () => {
+      const fetchMoreError = new Error('Failed to fetch records');
 
-    expect(finalResult).toEqual([mockPerson, mockPerson]);
+      mockFindManyRecordsLazy.mockResolvedValue({
+        data: {
+          people: {
+            edges: [{ node: { id: '1' }, cursor: '1' }],
+            pageInfo: {
+              hasNextPage: true,
+              hasPreviousPage: false,
+              startCursor: '1',
+              endCursor: '1',
+            },
+            totalCount: 2,
+          },
+        },
+        error: undefined,
+      });
+
+      mockFetchMoreRecordsLazy.mockResolvedValue({ error: fetchMoreError });
+
+      const { result } = renderHook(() =>
+        useLazyFetchAllRecords({
+          objectNameSingular: 'person',
+          limit: 1,
+        }),
+      );
+
+      await expect(async () => {
+        await act(async () => {
+          await result.current.fetchAllRecords();
+        });
+      }).rejects.toThrow('Failed to fetch records');
+
+      await waitFor(() => {
+        expect(result.current.isDownloading).toBe(false);
+        expect(result.current.progress).toEqual({ displayType: 'number' });
+      });
+    });
+
+    it('should reject when the initial page fetch fails', async () => {
+      const fetchError = new Error('Failed to fetch records');
+
+      mockFindManyRecordsLazy.mockResolvedValue({
+        data: null,
+        error: fetchError,
+      });
+
+      const { result } = renderHook(() =>
+        useLazyFetchAllRecords({
+          objectNameSingular: 'person',
+          limit: 1,
+        }),
+      );
+
+      await expect(async () => {
+        await act(async () => {
+          await result.current.fetchAllRecords();
+        });
+      }).rejects.toThrow('Failed to fetch records');
+
+      await waitFor(() => {
+        expect(result.current.isDownloading).toBe(false);
+        expect(result.current.progress).toEqual({ displayType: 'number' });
+      });
+    });
   });
 });
