+++ packages/twenty-server/src/engine/twenty-orm/utils/is-record-matching-rls-row-level-permission-predicate.util.ts
@@ -2,57 +2,24 @@
 
 import { isObject } from '@sniptt/guards';
 import {
-  FieldMetadataType,
-  type ActorFilter,
-  type AddressFilter,
   type AndObjectRecordFilter,
-  type ArrayFilter,
-  type BooleanFilter,
-  type CurrencyFilter,
-  type DateFilter,
-  type EmailsFilter,
-  type FloatFilter,
-  type FullNameFilter,
-  type IsFilter,
   type LeafObjectRecordFilter,
-  type LinksFilter,
-  type MultiSelectFilter,
   type NotObjectRecordFilter,
   type OrObjectRecordFilter,
-  type PhonesFilter,
-  type RatingFilter,
-  type RawJsonFilter,
   type RecordGqlOperationFilter,
-  type RichTextFilter,
-  type SelectFilter,
-  type StringFilter,
-  type TSVectorFilter,
-  type UUIDFilter,
-  type UUIDFilterValue,
 } from 'twenty-shared/types';
-import {
-  isDefined,
-  isEmptyObject,
-  isMatchingArrayFilter,
-  isMatchingBooleanFilter,
-  isMatchingCurrencyFilter,
-  isMatchingDateFilter,
-  isMatchingFloatFilter,
-  isMatchingMultiSelectFilter,
-  isMatchingRatingFilter,
-  isMatchingRawJsonFilter,
-  isMatchingRichTextFilter,
-  isMatchingSelectFilter,
-  isMatchingStringFilter,
-  isMatchingTSVectorFilter,
-  isMatchingUUIDFilter,
-} from 'twenty-shared/utils';
+import { isDefined, isEmptyObject } from 'twenty-shared/utils';
 
-import { computeMorphOrRelationFieldJoinColumnName } from 'src/engine/metadata-modules/field-metadata/utils/compute-morph-or-relation-field-join-column-name.util';
 import { getFlatFieldsFromFlatObjectMetadata } from 'src/engine/api/graphql/workspace-schema-builder/utils/get-flat-fields-for-flat-object-metadata.util';
 import { type FlatEntityMaps } from 'src/engine/metadata-modules/flat-entity/types/flat-entity-maps.type';
 import { type FlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type';
 import { type FlatObjectMetadata } from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata.type';
+import { evaluateRLSRowLevelPermissionLeafFields } from 'src/engine/twenty-orm/utils/evaluate-rls-row-level-permission-leaf-field.util';
+import {
+  combineAndResults,
+  combineOrResults,
+  type RLSPredicateEvaluation,
+} from 'src/engine/twenty-orm/utils/evaluate-rls-row-level-permission-predicate-evaluation.util';
 
 const isLeafFilter = (
   filter: RecordGqlOperationFilter,
@@ -75,106 +42,10 @@ const isNotFilter = (
   filter: RecordGqlOperationFilter,
 ): filter is NotObjectRecordFilter => 'not' in filter && !!filter.not;
 
-type RLSPredicateEvaluation = 'invalid' | boolean;
-
 const isRecordFilterObject = (
   filter: unknown,
 ): filter is RecordGqlOperationFilter =>
   isObject(filter) && !Array.isArray(filter);
-
-const combineAndResults = (
-  results: RLSPredicateEvaluation[],
-): RLSPredicateEvaluation => {
-  if (results.length === 0) {
-    return true;
-  }
-
-  if (results.some((result) => result === 'invalid')) {
-    return 'invalid';
-  }
-
-  return results.every((result) => result === true);
-};
-
-const combineOrResults = (
-  results: RLSPredicateEvaluation[],
-): RLSPredicateEvaluation => {
-  if (results.length === 0) {
-    return true;
-  }
-
-  if (results.some((result) => result === 'invalid')) {
-    return 'invalid';
-  }
-
-  return results.some((result) => result === true);
-};
-
-const evaluateFieldMatchResult = (
-  evaluation: () => boolean,
-): RLSPredicateEvaluation => {
-  try {
-    return evaluation();
-  } catch {
-    return 'invalid';
-  }
-};
-
-const hasOnlyAllowedDefinedFilterKeys = <T extends Record<string, unknown>>(
-  filter: T,
-  keys: readonly (keyof T)[],
-): boolean => {
-  const filterKeys = Object.keys(filter);
-
-  if (filterKeys.length === 0) {
-    return false;
-  }
-
-  return (
-    filterKeys.every((key) => keys.includes(key as keyof T)) &&
-    keys.some((key) => filter[key] !== undefined)
-  );
-};
-
-const evaluateCompositeSubFieldMatch = ({
-  subFieldFilter,
-  subFieldValue,
-}: {
-  subFieldFilter: StringFilter | RawJsonFilter | SelectFilter | UUIDFilter;
-  subFieldValue: unknown;
-}): boolean => {
-  if (
-    isDefined((subFieldFilter as RawJsonFilter).like) ||
-    isDefined((subFieldFilter as RawJsonFilter).is)
-  ) {
-    return isMatchingRawJsonFilter({
-      rawJsonFilter: subFieldFilter as RawJsonFilter,
-      value: subFieldValue as string,
-    });
-  }
-
-  if (isDefined((subFieldFilter as SelectFilter).eq)) {
-    return isMatchingSelectFilter({
-      selectFilter: subFieldFilter as SelectFilter,
-      value: subFieldValue as string,
-    });
-  }
-
-  if (
-    isDefined((subFieldFilter as UUIDFilter).eq) ||
-    isDefined((subFieldFilter as UUIDFilter).is)
-  ) {
-    return isMatchingUUIDFilter({
-      uuidFilter: subFieldFilter as UUIDFilter,
-      value: subFieldValue as UUIDFilterValue,
-    });
-  }
-
-  return isMatchingStringFilter({
-    stringFilter: subFieldFilter as StringFilter,
-    value: subFieldValue as string,
-  });
-};
 
 const evaluateRLSRowLevelPermissionPredicate = ({
   record,
@@ -306,341 +177,11 @@ const evaluateRLSRowLevelPermissionPredicate = ({
     flatFieldMetadataMaps,
   );
 
-  return combineAndResults(
-    Object.entries(filter).map(([filterKey, filterValue]) => {
-      if (!isDefined(filterValue)) {
-        return 'invalid';
-      }
-
-      if (isEmptyObject(filterValue)) {
-        return true;
-      }
-
-      const objectMetadataField =
-        objectFields.find((field) => field.name === filterKey) ??
-        objectFields.find(
-          (field) =>
-            (field.type === FieldMetadataType.RELATION ||
-              field.type === FieldMetadataType.MORPH_RELATION) &&
-            computeMorphOrRelationFieldJoinColumnName({ name: field.name }) ===
-              filterKey,
-        );
-
-      if (!isDefined(objectMetadataField)) {
-        return 'invalid';
-      }
-
-      const recordFieldValue = record[filterKey];
-
-      if (!isDefined(recordFieldValue)) {
-        if (isObject(filterValue)) {
-          return (filterValue as { is?: IsFilter })?.is === 'NULL';
-        }
-
-        return false;
-      }
-
-      switch (objectMetadataField.type) {
-        case FieldMetadataType.RATING:
-          return evaluateFieldMatchResult(() =>
-            isMatchingRatingFilter({
-              ratingFilter: filterValue as RatingFilter,
-              value: recordFieldValue,
-            }),
-          );
-        case FieldMetadataType.TEXT: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingStringFilter({
-              stringFilter: filterValue as StringFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.RICH_TEXT: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingRichTextFilter({
-              richTextFilter: filterValue as RichTextFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.SELECT:
-          return evaluateFieldMatchResult(() =>
-            isMatchingSelectFilter({
-              selectFilter: filterValue as SelectFilter,
-              value: recordFieldValue,
-            }),
-          );
-        case FieldMetadataType.MULTI_SELECT:
-          return evaluateFieldMatchResult(() =>
-            isMatchingMultiSelectFilter({
-              multiSelectFilter: filterValue as MultiSelectFilter,
-              value: recordFieldValue,
-            }),
-          );
-        case FieldMetadataType.ARRAY: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingArrayFilter({
-              arrayFilter: filterValue as ArrayFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.RAW_JSON: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingRawJsonFilter({
-              rawJsonFilter: filterValue as RawJsonFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.FULL_NAME: {
-          const fullNameFilter = filterValue as FullNameFilter;
-
-          if (
-            !hasOnlyAllowedDefinedFilterKeys(fullNameFilter, [
-              'firstName',
-              'lastName',
-            ])
-          ) {
-            return 'invalid';
-          }
-
-          return evaluateFieldMatchResult(
-            () =>
-              (fullNameFilter.firstName === undefined ||
-                isMatchingStringFilter({
-                  stringFilter: fullNameFilter.firstName,
-                  value: recordFieldValue.firstName,
-                })) &&
-              (fullNameFilter.lastName === undefined ||
-                isMatchingStringFilter({
-                  stringFilter: fullNameFilter.lastName,
-                  value: recordFieldValue.lastName,
-                })),
-          );
-        }
-        case FieldMetadataType.ADDRESS: {
-          const addressFilter = filterValue as AddressFilter;
-
-          const keys = [
-            'addressStreet1',
-            'addressStreet2',
-            'addressCity',
-            'addressState',
-            'addressCountry',
-            'addressPostcode',
-          ] as const;
-
-          if (!hasOnlyAllowedDefinedFilterKeys(addressFilter, keys)) {
-            return 'invalid';
-          }
-
-          return evaluateFieldMatchResult(() =>
-            keys.some((key) => {
-              const value = addressFilter[key];
-
-              if (value === undefined) {
-                return false;
-              }
-
-              return isMatchingStringFilter({
-                stringFilter: value,
-                value: recordFieldValue[key],
-              });
-            }),
-          );
-        }
-        case FieldMetadataType.LINKS: {
-          const linksFilter = filterValue as LinksFilter;
-
-          const keys = [
-            'primaryLinkLabel',
-            'primaryLinkUrl',
-            'secondaryLinks',
-          ] as const;
-
-          if (!hasOnlyAllowedDefinedFilterKeys(linksFilter, keys)) {
-            return 'invalid';
-          }
-
-          return evaluateFieldMatchResult(() =>
-            keys.some((key) => {
-              const value = linksFilter[key];
-
-              if (value === undefined) {
-                return false;
-              }
-
-              return evaluateCompositeSubFieldMatch({
-                subFieldFilter: value,
-                subFieldValue: recordFieldValue[key],
-              });
-            }),
-          );
-        }
-        case FieldMetadataType.DATE:
-        case FieldMetadataType.DATE_TIME: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingDateFilter({
-              dateFilter: filterValue as DateFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.NUMBER:
-        case FieldMetadataType.NUMERIC: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingFloatFilter({
-              floatFilter: filterValue as FloatFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.UUID: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingUUIDFilter({
-              uuidFilter: filterValue as UUIDFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.BOOLEAN: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingBooleanFilter({
-              booleanFilter: filterValue as BooleanFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.CURRENCY: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingCurrencyFilter({
-              currencyFilter: filterValue as CurrencyFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        case FieldMetadataType.ACTOR: {
-          const actorFilter = filterValue as ActorFilter;
-
-          const keys = ['source', 'name', 'workspaceMemberId'] as const;
-
-          if (!hasOnlyAllowedDefinedFilterKeys(actorFilter, keys)) {
-            return 'invalid';
-          }
-
-          return evaluateFieldMatchResult(() => {
-            if (isDefined(actorFilter.source)) {
-              return isMatchingSelectFilter({
-                selectFilter: actorFilter.source,
-                value: recordFieldValue.source,
-              });
-            }
-
-            if (isDefined(actorFilter.workspaceMemberId)) {
-              return isMatchingUUIDFilter({
-                uuidFilter: actorFilter.workspaceMemberId,
-                value: recordFieldValue.workspaceMemberId,
-              });
-            }
-
-            return (
-              actorFilter.name === undefined ||
-              isMatchingStringFilter({
-                stringFilter: actorFilter.name,
-                value: recordFieldValue.name,
-              })
-            );
-          });
-        }
-        case FieldMetadataType.EMAILS: {
-          const emailsFilter = filterValue as EmailsFilter;
-
-          const keys = ['primaryEmail', 'additionalEmails'] as const;
-
-          if (!hasOnlyAllowedDefinedFilterKeys(emailsFilter, keys)) {
-            return 'invalid';
-          }
-
-          return evaluateFieldMatchResult(() =>
-            keys.some((key) => {
-              const value = emailsFilter[key];
-
-              if (value === undefined) {
-                return false;
-              }
-
-              return evaluateCompositeSubFieldMatch({
-                subFieldFilter: value,
-                subFieldValue: recordFieldValue[key],
-              });
-            }),
-          );
-        }
-        case FieldMetadataType.PHONES: {
-          const phonesFilter = filterValue as PhonesFilter;
-
-          const keys = [
-            'primaryPhoneNumber',
-            'primaryPhoneCallingCode',
-            'additionalPhones',
-          ] as const;
-
-          if (!hasOnlyAllowedDefinedFilterKeys(phonesFilter, keys)) {
-            return 'invalid';
-          }
-
-          return evaluateFieldMatchResult(() =>
-            keys.some((key) => {
-              const value = phonesFilter[key];
-
-              if (value === undefined) {
-                return false;
-              }
-
-              return evaluateCompositeSubFieldMatch({
-                subFieldFilter: value,
-                subFieldValue: recordFieldValue[key],
-              });
-            }),
-          );
-        }
-        case FieldMetadataType.RELATION:
-        case FieldMetadataType.MORPH_RELATION: {
-          const isJoinColumn =
-            computeMorphOrRelationFieldJoinColumnName({
-              name: objectMetadataField.name,
-            }) === filterKey;
-
-          return evaluateFieldMatchResult(() => {
-            if (isJoinColumn) {
-              return isMatchingUUIDFilter({
-                uuidFilter: filterValue as UUIDFilter,
-                value: recordFieldValue,
-              });
-            }
-
-            return isMatchingUUIDFilter({
-              uuidFilter: filterValue as UUIDFilter,
-              value: recordFieldValue?.id ?? null,
-            });
-          });
-        }
-        case FieldMetadataType.TS_VECTOR: {
-          return evaluateFieldMatchResult(() =>
-            isMatchingTSVectorFilter({
-              tsVectorFilter: filterValue as TSVectorFilter,
-              value: recordFieldValue,
-            }),
-          );
-        }
-        default: {
-          return 'invalid';
-        }
-      }
-    }),
-  );
+  return evaluateRLSRowLevelPermissionLeafFields({
+    record,
+    filter,
+    objectFields,
+  });
 };
 
 export const isRecordMatchingRLSRowLevelPermissionPredicate = ({
